warnings--
uninitialized values are potentially being used in a number of places:

kexi/main/printing/kexisimpleprintingpagesetup.cpp:306 (objectId)
kexi/plugins/forms/widgets/kexidblineedit.cpp:123 (bottom, top)
filters/kpresenter/powerpoint/libppt/powerpoint.cpp:5470 (lastSize)
kexi/migration/importwizard.cpp:675 (cdata)
krita/plugins/filters/cimg/kis_cimg_filter.cc:471 (nn)
filters/krita/tiff/kis_tiff_converter.cc:304 (tiffReader)
filters/krita/tiff/kis_tiff_converter.cc:306 (postprocessor)

CCMAIL: koffice-devel@kde.org

svn path=/trunk/koffice/; revision=512034

+++ karbon/shapes/vellipse.cc
@@ -55,9 +55,9 @@ VEllipse::init()
 	// to radials
 	int nsegs;
 	if( m_startAngle < m_endAngle )
-	    nsegs = floor( ( m_endAngle - m_startAngle ) / 90.0 );
+	    nsegs = int( floor( ( m_endAngle - m_startAngle ) / 90.0 ) );
 	else
-	    nsegs = 4 - ceil( ( m_startAngle - m_endAngle ) / 90.0 );
+	    nsegs = 4 - int( ceil( ( m_startAngle - m_endAngle ) / 90.0 ) );
 	double startAngle = m_startAngle - 90.0;
 	if( startAngle < 0 ) startAngle += 360.0;
 	startAngle = VGlobal::pi_2 * ( startAngle / 90.0 );
